requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0